requirements.txt
openai
azure.identity
